octofit-tracker/frontend/octofit-tracker/backend/apps/leaderboard/urls.py
from rest_framework.routers import DefaultRouter
from .views import LeaderboardEntryViewSet

router = DefaultRouter()
router.register(r'leaderboard', LeaderboardEntryViewSet)

urlpatterns = router.urls
